Uploading Code that Demonstrates Stack Usage
This program reads information from four different text files, and assigns the information to four different stacks. The program then prints the read values. Then, the program merges two stacks into one, sorts the stack, and reprints the stack. When, the program prints the sorted stack, the stack is sorted/printed in the order of smallest to largest.